fix: Add upgrade pod as an extra deployment
In Syndesis 1.3.10 the upgrade pod was includede with help of an template
that gets created when the Syndesis template is instantiated.
This approach has the disadvantage that the overall template (the outter)
has a fixed name and to create this the old template would need
to be replaced.
But it is the update itself who should replace.
That catch-22 can be only resolved if the start of the update pod
is not defined within the template it's supposed to be replaced.

Therefore a new file 'fuse-ignite-upgrade.yml' will be created in `resources/`` .
This can be directly applied via `oc create -f fuse-ignite-upgrade.yml`
Therefor the documentation needs to be changed/updated.